+++ src/users/users.service.ts
@@ -36,6 +36,9 @@ export class UsersService {
 
   async remove(id: number) {
     const user = await this.userRepository.findOne(id);
+    if (!user) {
+      throw new NotFoundException(`User #${id} not found`);
+    }
     return this.userRepository.remove(user);
   }
 }
